Replace long by int in qi::atomic
Reviewed-on: http://gerrit.aldebaran.lan:8080/8416
Tested-by: jfreche <jfreche@aldebaran-robotics.com>

+++ qitype/details/signal.hxx
@@ -44,7 +44,7 @@ namespace qi
         return new BoostWeakPointerLock(weakPointer);
       }
     private:
-      qi::Atomic<long>  lockCount;
+      qi::Atomic<int>  lockCount;
       boost::shared_ptr<T> sharedPointer;
       boost::weak_ptr<T>   weakPointer;
     };
